deprecate InterleavedBuffer .setArray
